@@ -50,9 +50,9 @@ extend-exclude = "langtest/transform/constants.py"
 exclude = 'langtest/errors.py'
 
 [tool.poetry.dependencies]
-python = ">=3.8.1,<4.0"
-pydantic = "1.10.11"
-johnsnowlabs = { version = "4.3.5", optional = true }
+python = ">=3.9,<4.0"
+pydantic = "2.9.2"
+johnsnowlabs = {version = "5.5.0", optional = true}
 rouge-score = { version = "^0.1.2", optional = true }
 evaluate = { version = "^0.4.0", optional = true }
 transformers = "^4.44.2"
@@ -60,54 +60,58 @@ huggingface_hub = { version = ">0.16.0", optional = true}
 spacy = { version = ">=3.0.0", optional = true }
 nest-asyncio = "^1.5.0"
 openai = "^1.40.6"
-jsonlines = "^3.1.0"
+jsonlines = ">=3.1.0"
 torch = { version = "^2.0.0", optional = true }
 pandas = "^2.0.3"
 pyyaml = "^6.0"
-tqdm = "^4.65.0"
+tqdm = ">=4.65.0"
 cohere = { version = "^4.10.0", optional = true}
 ai21 = {version = "^1.1.0", optional = true}
 metaflow = {version = ">=2.9.0", optional = true}
-accelerate = {version = "0.33.0", optional = true}
+accelerate = {version = "0.34.2", optional = true}
 seqeval = {version = "^1.2.0", optional = true}
-mlflow = {version = "^2.16.2", optional = true}
+mlflow = {version = "^2.18.0", optional = true}
 datasets = {version = ">=2.14.0", optional = true}
 matplotlib = {version = "^3.7.2", optional = true}
 tenacity = {version = "^8.2.2", optional = true}
-langchain = {version = "^0.1.11", optional = true}
+langchain = {version = "0.3.6", optional = true}
 typing-extensions = "^4.10.0"
-langchain-openai = {version = "^0.1.4", optional = true}
+langchain-openai = {version = "0.2.5", optional = true}
 boto3 = {version = "^1.34.93", optional = true}
 importlib-resources = "^6.4.0"
-click = "^8.1.7"
-openpyxl = "^3.1.5"
-tables = "3.8.0"
-pillow = "10.0.0"
+click = ">=8.1.7"
+openpyxl = ">=3.1.5"
+tables = ">=3.8.0"
+pillow = ">=11.0.0"
+langchain-databricks = {version = "0.1.1", optional = true}
+langchain-community = {version = "0.3.5", optional = true}
 
 [tool.poetry.extras]
 transformers = ["transformers", "torch", "accelerate", "datasets"]
 evaluate = ["evaluate", "rouge-score", "seqeval"]
 spacy = ["spacy"]
-langchain = ["langchain", "langchain-openai"]
 johnsnowlabs = ["johnsnowlabs"]
-openai = ["openai", "langchain", "langchain-openai"]
-cohere = ["cohere", "langchain"]
-ai21 = ["ai21", "langchain"]
+openai = ["openai", "langchain", "langchain-openai", "langchain-community"]
+cohere = ["cohere", "langchain", "langchain-community"]
+ai21 = ["ai21", "langchain", "langchain-community"]
 huggingface_hub = ["huggingface_hub", "langchain"]
 metaflow = ["metaflow"]
 mlflow = ["mlflow"]
 datasets = ["datasets"]
 matplotlib = ["matplotlib"]
+databricks = ["langchain", "langchain-databricks", "langchain-community"]
 
 [tool.poetry.group.dev.dependencies]
 ipdb = "^0.13.13"
 pytest = "^7.4.0"
 taskipy = "^1.11.0"
-en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0.tar.gz" }
+en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0.tar.gz" }
 black = "^23.3.0"
 pyproject-flake8 = "<6.0.0"
 pre-commit = "^3.3.3"
 pydocstyle = "^6.3.0"
+ipykernel = "^6.29.5"
+pytest-cov = "^6.0.0"
 
 [tool.taskipy.tasks]
 test = "pytest tests/* -v --disable-warnings"
