@@ -1,6 +1,7 @@
 import re
 import string
-from typing import List, Optional, Tuple
+from textwrap import dedent
+from typing import List, Mapping, Optional, Tuple
 from ..utils.custom_types.helpers import HashableDict
 
 template = """You are a teacher grading a quiz.
@@ -22,10 +23,82 @@ server_prompt = "Perform the task to the best of your ability."
 input_variables = ["query", "result", "answer"]
 
 
+class EvalTemplate:
+    """
+    The EvalTemplate class provides a method to build a prompt for evaluating student answers
+    based on a given rubric. The prompt is designed for a teacher to grade a quiz by comparing
+    the student's answer with the true answer and scoring it according to specified criteria.
+
+    Methods
+    -------
+    build_prompt(rubic_score: Mapping[str, str] = {"CORRECT": None, "INCORRECT": None}) -> str
+        Constructs and returns a grading prompt based on the provided rubric scores.
+
+    """
+
+    @staticmethod
+    def build_prompt(
+        rubic_score: Mapping[str, str] = {
+            "CORRECT": None,
+            "INCORRECT": None,
+        }
+    ):
+        """ """
+        grade_list = list(rubic_score.keys())
+        grade_list = ", ".join(grade_list[:-1]) + f" or {grade_list[-1]}"
+
+        eval_criteria = [
+            f"{grade_name}: {criteria}\n"
+            for grade_name, criteria in rubic_score.items()
+            if criteria
+        ]
+        prompt = (
+            "You are a teacher grading a quiz. You are given a question, the student's "
+            "answer, and the true answer, and are asked to score the student answer as either "
+            f"{grade_list}."
+        )
+
+        if eval_criteria:
+            eval_criteria = "".join(eval_criteria)
+            prompt += dedent(
+                f"""\n\nScore the student answer based on the following criteria:\n{eval_criteria}"""
+            )
+
+        prompt += dedent(
+            f"""
+        Example Format:
+        QUESTION: question here
+        STUDENT ANSWER: student's answer here
+        TRUE ANSWER: true answer here
+        GRADE: {grade_list} here
+
+        {
+            ("Grade the student answers based ONLY on their factual accuracy. Ignore differences"
+             " in punctuation and phrasing between the student answer and true answer. It is OK "
+             "if the student answer contains more or relevant information than the true answer, as"
+             " long as it does not contain any conflicting statements. Begin!")
+        }
+
+        QUESTION: {{query}}
+        STUDENT ANSWER: {{result}}
+        TRUE ANSWER: {{answer}}
+        GRADE:"""
+        )
+        return prompt
+
+
 class LlmEval:
     """llm_eval for evaluating question answering."""
 
-    def __init__(self, llm, template=template, input_variables=input_variables):
+    grade_list = None
+
+    def __init__(
+        self,
+        llm,
+        template=template,
+        input_variables=input_variables,
+        grade_list=None,
+    ):
         """
         Initializes the LlmEval object.
 
@@ -42,6 +115,7 @@ class LlmEval:
         self.template = template
         self.input_variables = input_variables
         self.server_prompt = server_prompt
+        LlmEval.grade_list = grade_list
 
         expected_input_vars = {"query", "answer", "result"}
         if expected_input_vars != set(self.input_variables):
@@ -52,33 +126,55 @@ class LlmEval:
 
     @staticmethod
     def _get_score(text: str) -> Optional[Tuple[str, int]]:
-        match = re.search(r"grade:\s*(correct|incorrect)", text.strip(), re.IGNORECASE)
+        if LlmEval.grade_list is None:
+            default_grades = ["CORRECT", "INCORRECT"]
+            grade_list_pattern = f"grade:\\s*({'|'.join(default_grades).lower()})"
+        else:
+            grade_list_pattern = f"(?:grade\\s*)?({'|'.join(LlmEval.grade_list).lower()})"
+
+        match = re.search(grade_list_pattern, text.strip(), re.IGNORECASE)
         if match:
-            if match.group(1).upper() == "CORRECT":
-                return "CORRECT", 1
-            elif match.group(1).upper() == "INCORRECT":
-                return "INCORRECT", 0
-        try:
-            first_word = (
-                text.strip()
-                .split()[0]
-                .translate(str.maketrans("", "", string.punctuation))
-            )
-            if first_word.upper() == "CORRECT":
-                return "CORRECT", 1
-            elif first_word.upper() == "INCORRECT":
-                return "INCORRECT", 0
-            last_word = (
-                text.strip()
-                .split()[-1]
-                .translate(str.maketrans("", "", string.punctuation))
-            )
-            if last_word.upper() == "CORRECT":
-                return "CORRECT", 1
-            elif last_word.upper() == "INCORRECT":
-                return "INCORRECT", 0
-        except IndexError:
-            pass
+            grade = match.group(1).upper()
+            if LlmEval.grade_list is None:
+                if grade == "CORRECT":
+                    return "CORRECT", 1
+                elif grade == "INCORRECT":
+                    return "INCORRECT", 0
+            elif grade in LlmEval.grade_list:
+                return grade, LlmEval.grade_list.index(grade)
+        else:
+            try:
+                # Check for first word
+                first_word = (
+                    text.strip()
+                    .split()[0]
+                    .translate(str.maketrans("", "", string.punctuation))
+                )
+                if LlmEval.grade_list is None:
+                    if first_word.upper() == "CORRECT":
+                        return "CORRECT", 1
+                    elif first_word.upper() == "INCORRECT":
+                        return "INCORRECT", 0
+                elif first_word.upper() in LlmEval.grade_list:
+                    return first_word.upper(), LlmEval.grade_list.index(
+                        first_word.upper()
+                    )
+
+                # Check for last word
+                last_word = (
+                    text.strip()
+                    .split()[-1]
+                    .translate(str.maketrans("", "", string.punctuation))
+                )
+                if LlmEval.grade_list is None:
+                    if last_word.upper() == "CORRECT":
+                        return "CORRECT", 1
+                    elif last_word.upper() == "INCORRECT":
+                        return "INCORRECT", 0
+                elif last_word.upper() in LlmEval.grade_list:
+                    return last_word.upper(), LlmEval.grade_list.index(last_word.upper())
+            except IndexError:
+                pass
         return None
 
     @staticmethod
