@@ -4,7 +4,7 @@ import random
 from langtest.logger import logger as logging
 import re
 from abc import ABC, abstractmethod
-from typing import Dict, List
+from typing import Dict, List, TypedDict
 
 from langtest.modelhandler.modelhandler import ModelAPI
 from langtest.transform.base import ITests, TestFactory
@@ -210,7 +210,7 @@ class BiasTestFactory(ITests):
         return all_samples
 
     @staticmethod
-    def available_tests() -> Dict:
+    def available_tests() -> Dict[str, type["BaseBias"]]:
         """
         Get a dictionary of all available tests, with their names as keys and their corresponding classes as values.
 
@@ -246,6 +246,9 @@ class BaseBias(ABC):
         "question-answering",
         "summarization",
     ]
+
+    # Config Hint for the bias tests
+    TestConfig = TypedDict("TestConfig", min_pass_rate=float)
 
     @abstractmethod
     def transform(self, sample_list: List[Sample], *args, **kwargs) -> List[Sample]:
